Fixing flake8 issues
