+++ pandas/io/formats/common.py
@@ -6,6 +6,7 @@ Common helper methods used in different submodules of pandas.io.formats
 from pandas.core.dtypes.common import is_integer
 from pandas import compat
 
+
 def get_level_lengths(levels, sentinel=''):
     """For each index in each level the function returns lengths of indexes.
 
@@ -45,32 +46,15 @@ def get_level_lengths(levels, sentinel=''):
 
     return result
 
+
 def _put_lines(buf, lines):
     if any(isinstance(x, compat.text_type) for x in lines):
         lines = [compat.text_type(x) for x in lines]
     buf.write('\n'.join(lines))
 
 
-def _binify(cols, line_width):
-    adjoin_width = 1
-    bins = []
-    curr_width = 0
-    i_last_column = len(cols) - 1
-    for i, w in enumerate(cols):
-        w_adjoined = w + adjoin_width
-        curr_width += w_adjoined
-        if i_last_column == i:
-            wrap = curr_width + 1 > line_width and i > 0
-        else:
-            wrap = curr_width + 2 > line_width and i > 0
-        if wrap:
-            bins.append(i)
-            curr_width = w_adjoined
-
-    bins.append(len(cols))
-    return bins
-
 class TableFormatter(object):
+
     is_truncated = False
     show_dimensions = None
